+++ packages/code/src/__stories__/utils.ts
@@ -1,0 +1,12 @@
+export const getRange = (
+  lineNumber: number,
+  startColumn: number,
+  endColumn: number,
+) => {
+  return {
+    startLineNumber: lineNumber,
+    endLineNumber: lineNumber,
+    startColumn: startColumn,
+    endColumn: endColumn,
+  };
+};