+++ packages/code/src/index.ts
@@ -1,0 +1,4 @@
+export * from './SmartEditor';
+export * from './hooks';
+export * from './api';
+export * from './themes';