+++ packages/code/vite.config.ts
@@ -1,0 +1,56 @@
+import { resolve } from 'path';
+import { defineConfig } from 'vite';
+import { visualizer } from 'rollup-plugin-visualizer';
+// Used from @storybook/react-vite dependencies
+import react from '@vitejs/plugin-react';
+
+/** @type {import('vite').UserConfig} */
+export default defineConfig({
+  build: {
+    sourcemap: true,
+    lib: {
+      entry: resolve(__dirname, 'src/index.ts'),
+      name: 'index',
+      fileName: 'index',
+    },
+    rollupOptions: {
+      external: [
+        'react',
+        'react-dom',
+        'monaco-editor-core',
+        '@devoinc/genesys-ui',
+        '@devoinc/genesys-brand-devo',
+        'styled-components',
+      ],
+      output: {
+        globals: {
+          react: 'React',
+          'react-dom': 'reactDom',
+          'styled-components': 'styled',
+          'monaco-editor-core': 'monacoEditorCore',
+          '@devoinc/genesys-ui': 'genesysUi',
+        },
+      },
+    },
+  },
+  plugins: [
+    react({
+      babel: {
+        plugins: [
+          [
+            'babel-plugin-styled-components',
+            {
+              ssr: false,
+              pure: true,
+              displayName: true,
+              fileName: false,
+            },
+          ],
+        ],
+      },
+    }),
+    visualizer({
+      filename: 'dist/stats.html',
+    }),
+  ],
+});